feat(partners): Bartending Partners tab — 33 researched + fully enriched Austin mobile bartending companies
- Refactor prospect table into generic PartnerProspectsView (config-driven
  labels, CSV category, master outreach template); StrPartnersView and new
  BartendingPartnersView are thin wrappers
- New 'Bartending Partners' tab (?tab=bartenders): 33 Austin mobile-
  bartending / mobile-bar / cocktail-catering companies with contact,
  socials, scraped logos (31/33), and FULL enrichment on every row —
  management, offering, business model, reputation, partnership angles,
  personalized outreach email
- 6 rows matched to existing BARTENDER affiliates (DTR, PourTwentyFour,
  Social Butterfly, Kickstand, Tap Truck, Silvercloud) — marked created
  with their live partner pages; their emails are activation-style
- Bartender-specific master outreach template (dry-hire supply-chain
  pitch: TX mobile bartenders can't sell liquor — POD closes the gap)
- Trails End Mobile Bar researched then disqualified (Iowa-based)
