+++ src/app/admin/brians-stuff/BriansStuffTabs.tsx
@@ -30,7 +30,8 @@ type TabKey =
   | 'pathways'
   | 'logic'
   | 'docs'
-  | 'str';
+  | 'str'
+  | 'bartenders';
 
 export default function BriansStuffTabs({
   playbook,
@@ -44,6 +45,7 @@ export default function BriansStuffTabs({
   logic,
   docs,
   str,
+  bartenders,
   initialTab = 'playbook',
 }: {
   playbook: ReactNode;
@@ -57,6 +59,7 @@ export default function BriansStuffTabs({
   logic: ReactNode;
   docs: ReactNode;
   str: ReactNode;
+  bartenders: ReactNode;
   initialTab?: TabKey;
 }) {
   const [tab, setTab] = useState<TabKey>(initialTab);
@@ -110,6 +113,9 @@ export default function BriansStuffTabs({
           <TabButton active={tab === 'str'} onClick={() => setTab('str')}>
             🏡 STR Partners
           </TabButton>
+          <TabButton active={tab === 'bartenders'} onClick={() => setTab('bartenders')}>
+            🍸 Bartending Partners
+          </TabButton>
         </div>
       </div>
 
@@ -148,6 +154,9 @@ export default function BriansStuffTabs({
       </div>
       <div hidden={tab !== 'str'} className="px-6 md:px-10 py-8">
         <TabErrorBoundary tabName="STR Partners">{str}</TabErrorBoundary>
+      </div>
+      <div hidden={tab !== 'bartenders'} className="px-6 md:px-10 py-8">
+        <TabErrorBoundary tabName="Bartending Partners">{bartenders}</TabErrorBoundary>
       </div>
     </div>
   );
