+++ src/app/admin/brians-stuff/page.tsx
@@ -10,6 +10,7 @@ import OrderPathwaysView from '@/components/admin/OrderPathwaysView';
 import RecommendationLogicView from '@/components/admin/RecommendationLogicView';
 import EnrichmentDocsView from '@/components/admin/EnrichmentDocsView';
 import StrPartnersView from '@/components/admin/StrPartnersView';
+import BartendingPartnersView from '@/components/admin/BartendingPartnersView';
 import BriansStuffTabs from './BriansStuffTabs';
 
 export const metadata: Metadata = {
@@ -44,6 +45,7 @@ export default async function Page({ searchParams }: { searchParams: SP }) {
     'logic',
     'docs',
     'str',
+    'bartenders',
   ] as const;
   const initialTab = (
     (VALID_TABS as readonly string[]).includes(params.tab ?? '')
@@ -65,6 +67,7 @@ export default async function Page({ searchParams }: { searchParams: SP }) {
       logic={<RecommendationLogicView />}
       docs={<EnrichmentDocsView />}
       str={<StrPartnersView />}
+      bartenders={<BartendingPartnersView />}
     />
   );
 }
